Added Authorization for edit and delete post

+++ social-network-app/expressjs-rest-api/controllers/feed.js
@@ -135,6 +135,13 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not Authorized");
+        error.statusCode = 403;
+        throw error;
+      }
+
       if (imageUrl !== post.imageUrl) {
         deleteImage(post.imageUrl);
       }
@@ -166,6 +173,13 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not Authorized");
+        error.statusCode = 403;
+        throw error;
+      }
+
       deleteImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
     })
